Move type juggling to EditorLocationSearch from PostMetadata.
Previously, I was converting undefined to '' in PostMetadata when
retrieving the geo_label meta value.  However, doing that conversion
there has two downsides:

1) It makes geoLabel() less consistent with the other methods in that
file.  The reason for that inconsistency isn't readily apparently when
viewing that file in isolation.

2) That type conversion is really only needed because of a quirk in the
Search component.  If someone was to use the geoLabel() method in a
different context, they might think the conversion was irrelevant and
drop it without realizing the side effect.

Doing the conversion to an empty string in EditorLocationSearch instead
allows us to explain the specific context driving that change -- namely,
that <Search> treats undefined and '' differently in its
componentWillReceiveProps() method.